Standardize tool footer and refactor AI Hallucination Spotter rule database
