+++ index.html
@@ -2178,7 +2178,7 @@
         <li><a href="#projects">Projects</a></li>
         <li><a href="#about">About</a></li>
         <li><a href="#contact">Contact</a></li>
-        <li><a href="tutorials/index.html">Tutorials</a></li>
+        <!-- <li><a href="tutorials/index.html">Tutorials</a></li> -->
         <div class="nav-social">
           <a href="https://www.linkedin.com/in/hen-faibish-0469931b0/" target="_blank" rel="noopener noreferrer"
             title="LinkedIn">
