+++ tools/ai-hallucination-spotter/rules/index.js
@@ -1,0 +1,17 @@
+import { pythonRules } from './python.js';
+import { javascriptRules } from './javascript.js';
+import { javaRules } from './java.js';
+import { typescriptRules } from './typescript.js';
+import { goRules } from './go.js';
+import { rustRules } from './rust.js';
+import { csharpRules } from './csharp.js';
+
+export const superRules = {
+    python: pythonRules,
+    javascript: javascriptRules,
+    java: javaRules,
+    typescript: typescriptRules,
+    go: goRules,
+    rust: rustRules,
+    csharp: csharpRules
+};
